refactor: Update group chat message model to use int64 for IDs
Updated the GroupChatMessage model to use int64 for the ID, GroupID, and SenderID fields. This change aligns the data types with the database schema and ensures consistency across the application, preventing potential data truncation or overflow issues.

+++ backend/internal/models/group_chat_message.go
@@ -3,9 +3,9 @@ package models
 import "time"
 
 type GroupChatMessage struct {
-	ID        int       `json:"id"`
-	GroupID   int       `json:"group_id"`
-	SenderID  int       `json:"sender_id"`
+	ID        int64     `json:"id"`
+	GroupID   int64     `json:"group_id"`
+	SenderID  int64     `json:"sender_id"`
 	Content   string    `json:"content"`
 	CreatedAt time.Time `json:"created_at"`
 }
